[DEV] App to deal with remote desktops (pre/post)

+++ portal/experimentstool/views.py
@@ -40,12 +40,13 @@ def experimentstool(request):
 
 @login_required
 def get_hpc_list(request):
-    try:
+    response = _get_hpc_list(request.user)
+    if response:
         hpc_list = serializers.serialize(
             'json',
-            HPCInfrastructure.objects.filter(owner=request.user)
+            response
         )
-    except HPCInfrastructure.DoesNotExist:
+    else:
         hpc_list = []
     return JsonResponse(hpc_list, safe=False)
 
@@ -462,6 +463,14 @@ def remove_blueprint(request):
     return JsonResponse(_remove_blueprint(blueprint))
 
 
+def _get_hpc_list(user):
+    try:
+        hpc_list = HPCInfrastructure.objects.filter(owner=user)
+    except HPCInfrastructure.DoesNotExist:
+        hpc_list = []
+    return hpc_list
+
+
 def _add_hpc(name, owner, host, user, password, tz, manager):
     hpc = HPCInfrastructure.objects.create(name=name,
                                            owner=owner,
@@ -474,15 +483,23 @@ def _add_hpc(name, owner, host, user, password, tz, manager):
 
 
 def _delete_hpc(owner, pk):
+    hpc = _get_hpc_infrastructure(pk)
+    if not hpc:
+        return {'error': 'HPC does not exists'}
+
+    if (owner == hpc.owner):
+        hpc.delete()
+        return {'hpc': model_to_dict(hpc)}
+    else:
+        return {'error': 'HPC does not belong to user'}
+
+
+def _get_hpc_infrastructure(pk):
     try:
         hpc = HPCInfrastructure.objects.get(pk=pk)
-        if (owner == hpc.owner):
-            hpc.delete()
-            return {'hpc': model_to_dict(hpc)}
-        else:
-            return {'error': 'HPC does not belong to user'}
     except HPCInfrastructure.DoesNotExist:
-        return {'error': 'HPC does not exists'}
+        hpc = None
+    return hpc
 
 
 def _get_dataset(dataset_name):
